Test release build of bdk 0.29
