+++ keys/Cargo.toml
@@ -29,7 +29,8 @@ redgold-schema = { path = "../schema", version = "0.1.22"}
 
 # Make all the non-WASM dependencies optional
 tokio = { version = "1.36.0", features = ["rt"]}
-bdk = {package="bdk-debug", version = "0.28.1-debug", features = ["default", "keys-bip39", "bip39"]}
+#bdk = {package="bdk-debug", version = "0.28.1-debug", features = ["default", "keys-bip39", "bip39", "hardware-signer"]}
+bdk = {package="bdk", version = "0.29.0", features = ["default", "keys-bip39", "bip39", "hardware-signer"]}
 rust-crypto = { version = "0.2.36"}
 hex = {workspace = true}
 serde = {workspace = true}
