+++ src/gui/tabs/mod.rs
@@ -6,5 +6,6 @@ pub mod server_tab;
 pub mod address_tab;
 pub mod otp_tab;
 pub mod recovery_tab;
+pub mod keygen_subtab;
 pub mod keys_tab;
